feat(User)添加用户增加、删除功能 refactor(control)添加功能优化
通过获取用户对象添加用户、删除用户，添加功能直接调用mybatis plus提供方法

+++ src/main/java/com/cucaqi/service/impl/LesseeServiceImpl.java
@@ -2,6 +2,7 @@ package com.cucaqi.service.impl;
 
 import com.cucaqi.entity.Lessee;
 import com.cucaqi.mapper.LesseeMapper;
+import com.cucaqi.mapper.UserMapper;
 import com.cucaqi.service.ILesseeService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,6 +22,8 @@ import java.util.List;
 public class LesseeServiceImpl extends ServiceImpl<LesseeMapper, Lessee> implements ILesseeService {
     @Autowired
     private LesseeMapper lesseeMapper;
+    @Autowired
+    private UserMapper userMapper;
     @Override
     /**
      * 获取所有租户信息
@@ -28,12 +31,31 @@ public class LesseeServiceImpl extends ServiceImpl<LesseeMapper, Lessee> impleme
     public List<Lessee> getLesseeList(){
         return lesseeMapper.getLesseeList();
     }
-@Override
+    @Override
     public Integer addLessee(Lessee lessee){
         return lesseeMapper.addLessee(lessee);
     }
+
+    /**
+     * 删除租户
+     * @param id
+     * @return
+     */
     @Override
     public Integer deleteLessee(int id){
-        return  lesseeMapper.deleteLessee(id);
+
+        Integer count = userMapper.tellLessee(id);
+        //判断用户表中是否有关联的该租户
+        if(count==0){
+            //无关联租户，即可删除该租户
+            return  lesseeMapper.deleteLessee(id);
+        }
+        else{
+            //有关联租户，即不可删除该租户
+            return -100;
+        }
+
+
+
     }
 }
